Move LAST_SAVED_FILENAME to FileUtils
