+++ collect_app/src/main/java/org/odk/collect/android/upload/InstanceGoogleSheetsUploader.java
@@ -65,7 +65,7 @@ import java.util.regex.Pattern;
 import timber.log.Timber;
 
 import static org.odk.collect.android.logic.FormController.INSTANCE_ID;
-import static org.odk.collect.android.logic.FormController.LAST_SAVED_FILENAME;
+import static org.odk.collect.android.utilities.FileUtils.LAST_SAVED_FILENAME;
 
 public class InstanceGoogleSheetsUploader extends InstanceUploader {
     private static final String PARENT_KEY = "PARENT_KEY";
